requirements.txt
fastapi
uvicorn[standard]
pydantic
httpx
PyMuPDF
google-generativeai
python-dotenv
sqlmodel
aiosqlite